requirements.txt
streamlit==1.51.0
pandas
markitdown[pdf]
pypdf
pdfminer.six
google-generativeai
python-dotenv
PyMuPDF